requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
plotly==5.17.0
pillow==10.0.0
streamlit-option-menu==0.3.6
pymongo==4.6.1
passlib==1.7.4
python-dotenv==1.0.0 